Move the transloco services & config to the common directory.
