+++ apps/client/src/app/common/transloco/transloco-loader.ts
@@ -1,0 +1,13 @@
+import { HttpClient } from "@angular/common/http";
+import { inject, Injectable } from "@angular/core";
+import { Translation, TranslocoLoader } from "@ngneat/transloco";
+
+@Injectable({ providedIn: 'root' })
+export class TranslocoHttpLoader implements TranslocoLoader {
+  private http = inject(HttpClient);
+
+  getTranslation(lang: string) {
+    return this.http.get<Translation>(`/assets/i18n/${ lang }.json`);
+  }
+}
+
